requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Django==3.1.5
djangorestframework==3.12.2
Flask==1.1.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.6.0
numpy==1.19.5
pylint==2.6.0
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
wikipedia==1.4.0
wolframalpha==4.1.1
wrapt==1.12.1
xmltodict==0.12.0
gunicorn==20.0.4
imageai==2.1.6